+++ spec/middleware/request_stats_spec.rb
@@ -9,15 +9,6 @@ describe Napa::Middleware::RequestStats do
     ENV['STATSD_PORT'] = '8125'
   end
 
-  it 'should increment api_requests counter' do
-    expect(Napa::Stats.emitter).to receive(:increment).with('request_count')
-    expect(Napa::Stats.emitter).to receive(:increment).with('path.get.test.path.request_count')
-    app = lambda { |env| [200, { 'Content-Type' => 'application/json' }, Array.new] }
-    middleware = Napa::Middleware::RequestStats.new(app)
-    env = Rack::MockRequest.env_for('/test/path')
-    middleware.call(env)
-  end
-
   it 'should send the api_response_time' do
     expect(Napa::Stats.emitter).to receive(:timing).with('response_time', an_instance_of(Float))
     expect(Napa::Stats.emitter).to receive(:timing).with('path.get.test.path.response_time', an_instance_of(Float))